requirements.txt
Cython

opencv-python>=4.1.1

numpy==1.19.4
pandas==1.1.5
matplotlib==3.3.4

PyYAML>=5.3
scipy==1.4.1
tensorboard>=2.2
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
seaborn
scikit-learn
scikit-image

paho-mqtt>=1.5.1
